Add script to build libight for Android
In the files `config/api.list` and `config/arch.list` it is
possible to specify the API levels and the architectures of
the target devices.

Launching the script `build.sh`, the toolchains for the APIs and
architectures specified are created, (using the NDK tool
'make-standalone-toolchain.sh') and libight is built for the same
configurations.